refactor: 레디스 연결 모듈 파이프라인 -> 코어로 변경
