+++ ET-core/src/main/java/com/team1/etcore/EtCoreApplication.java
@@ -4,7 +4,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-@SpringBootApplication(scanBasePackages = {"com.team1.etcommon"})
+@SpringBootApplication(scanBasePackages = {"com.team1.etcommon","com.team1.etcore"})
 @EnableJpaAuditing
 public class EtCoreApplication {
