+++ ET-core/src/main/java/com/team1/etcore/chart/ChartDataController.java
@@ -1,6 +1,6 @@
-package com.team1.etpipeline.redis;
+package com.team1.etcore.chart;
 
-import com.team1.etpipeline.redis.dto.StockResponseDto;
+import com.team1.etcore.chart.dto.StockResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/stocks")
+@RequestMapping("/api/charts")
 @RequiredArgsConstructor
 public class ChartDataController {
     private final ChartDataService chartDataService;
