+++ ET-core/src/main/java/com/team1/etcore/chart/ChartDataRepository.java
@@ -1,12 +1,11 @@
-package com.team1.etpipeline.redis;
+package com.team1.etcore.chart;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.team1.etpipeline.redis.dto.StockResponseDto;
+import com.team1.etcore.chart.dto.StockResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Repository;
-import org.springframework.stereotype.Service;
 
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
