+++ build.gradle
@@ -131,6 +131,7 @@ project(':ET-core') {
         implementation project(':ET-common')
         implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
         implementation 'org.springframework.boot:spring-boot-starter-web'
+        implementation 'org.springframework.boot:spring-boot-starter-data-redis'
         implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
         runtimeOnly 'com.mysql:mysql-connector-j'
         testImplementation 'org.springframework.boot:spring-boot-starter-test'
@@ -172,7 +173,6 @@ project(':ET-arcade') {
 project(':ET-pipeline') {
     dependencies {
         implementation 'org.springframework.boot:spring-boot-starter-web'
-        implementation 'org.springframework.boot:spring-boot-starter-data-redis'
         implementation 'org.springframework.kafka:spring-kafka'
         testImplementation 'org.springframework.boot:spring-boot-starter-test'
         testImplementation 'org.springframework.kafka:spring-kafka-test'
